multi-agent-teacher/src/config.py
import os
from dotenv import load_dotenv
from pathlib import Path

# Load environment variables from .env if present
BASE_DIR = Path(__file__).resolve().parent.parent
ENV_PATH = BASE_DIR / ".env"
if ENV_PATH.exists():
    load_dotenv(ENV_PATH)

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")

DATA_DIR = BASE_DIR / "data"
CORPUS_DIR = DATA_DIR / "corpus"

EMBEDDING_MODEL = "text-embedding-3-small"
CHAT_MODEL = "gpt-4o-mini"

TOP_K_SNIPPETS = 3
