multi-agent-teacher/requirements.txt
openai>=1.6.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.0.0
jupyter>=1.0.0
matplotlib>=3.8.0
